:recycle: api: move scenario simulator routes from lib to src
