+++ api/lib/infrastructure/repositories/certification-assessment-repository.js
@@ -3,7 +3,7 @@ import { DomainTransaction } from '../DomainTransaction.js';
 import { CertificationAssessment } from '../../domain/models/CertificationAssessment.js';
 import { CertificationChallengeWithType } from '../../domain/models/CertificationChallengeWithType.js';
 import { Answer } from '../../../src/evaluation/domain/models/Answer.js';
-import * as challengeRepository from './challenge-repository.js';
+import * as challengeRepository from '../../../src/certification/shared/infrastructure/repositories/challenge-repository.js';
 import * as answerStatusDatabaseAdapter from '../adapters/answer-status-database-adapter.js';
 import { knex } from '../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../domain/errors.js';
