+++ api/lib/routes.js
@@ -35,7 +35,6 @@ import * as organizationInvitations from './application/organization-invitations
 import * as organizations from './application/organizations/index.js';
 import * as organizationsAdministration from './application/organizations-administration/index.js';
 import * as organizationLearners from './application/organization-learners/index.js';
-
 import * as passwords from './application/passwords/index.js';
 import * as poleEmploi from './application/pole-emploi/index.js';
 import * as prescribers from './application/prescribers/index.js';
@@ -43,7 +42,6 @@ import * as progressions from './application/progressions/index.js';
 import * as saml from './application/saml/index.js';
 import * as stageCollection from './application/stage-collections/index.js';
 import * as scoringSimulator from './application/scoring-simulator/index.js';
-import * as scenarioSimulator from './application/scenarios-simulator/index.js';
 import * as scorecards from './application/scorecards/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
 import * as supOrganizationLearners from './application/sup-organization-learners/index.js';
@@ -101,7 +99,6 @@ const routes = [
   prescribers,
   progressions,
   saml,
-  scenarioSimulator,
   scoringSimulator,
   organizationInvitations,
   scorecards,
