+++ api/scripts/certification/trigger-autojury-on-finalized-sessions.js
@@ -4,7 +4,7 @@ import { handleAutoJury } from '../../lib/domain/events/handle-auto-jury.js';
 import * as certificationIssueReportRepository from '../../src/certification/shared/infrastructure/repositories/certification-issue-report-repository.js';
 import * as certificationAssessmentRepository from '../../lib/infrastructure/repositories/certification-assessment-repository.js';
 import * as certificationCourseRepository from '../../lib/infrastructure/repositories/certification-course-repository.js';
-import * as challengeRepository from '../../lib/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../../src/certification/shared/infrastructure/repositories/challenge-repository.js';
 import { logger } from '../../lib/infrastructure/logger.js';
 import { SessionFinalized } from '../../lib/domain/events/SessionFinalized.js';
 import { eventDispatcher } from '../../lib/domain/events/index.js';
