+++ api/server.js
@@ -23,6 +23,7 @@ import {
 import { learnerManagementRoutes } from './src/prescription/learner-management/routes.js';
 import { devcompRoutes } from './src/devcomp/routes.js';
 import { schoolRoutes } from './src/school/routes.js';
+import { scenarioSimulatorRoutes } from './src/certification/flash-certification/routes.js';
 
 monitoringTools.installHapiHook();
 
@@ -119,6 +120,7 @@ const setupRoutesAndPlugins = async function (server) {
     routes,
     evaluationRoutes,
     certificationSessionRoutes,
+    scenarioSimulatorRoutes,
     attachTargetProfileRoutes,
     complementaryCertificationRoutes,
     devcompRoutes,
