+++ api/src/certification/flash-certification/application/scenario-simulator-controller.js
@@ -1,13 +1,13 @@
-import { usecases } from '../../domain/usecases/index.js';
-import { extractLocaleFromRequest } from '../../infrastructure/utils/request-response-utils.js';
-import { random } from '../../infrastructure/utils/random.js';
-import { scenarioSimulatorBatchSerializer } from '../../infrastructure/serializers/jsonapi/scenario-simulator-batch-serializer.js';
-import { parseCsv } from '../../../scripts/helpers/csvHelpers.js';
-import { pickAnswerStatusService } from '../../domain/services/pick-answer-status-service.js';
-import { HttpErrors } from '../http-errors.js';
+import { usecases } from '../../shared/domain/usecases/index.js';
+import { extractLocaleFromRequest } from '../../../../lib/infrastructure/utils/request-response-utils.js';
+import { random } from '../../../../lib/infrastructure/utils/random.js';
+import { scenarioSimulatorBatchSerializer } from '../../../../lib/infrastructure/serializers/jsonapi/scenario-simulator-batch-serializer.js';
+import { parseCsv } from '../../../../scripts/helpers/csvHelpers.js';
+import { pickAnswerStatusService } from '../../../../lib/domain/services/pick-answer-status-service.js';
+import { HttpErrors } from '../../../../lib/application/http-errors.js';
 import _ from 'lodash';
-import { pickChallengeService } from '../../domain/services/pick-challenge-service.js';
-import { FlashAssessmentSuccessRateHandler } from '../../domain/models/FlashAssessmentSuccessRateHandler.js';
+import { pickChallengeService } from '../../../../lib/domain/services/pick-challenge-service.js';
+import { FlashAssessmentSuccessRateHandler } from '../../../../lib/domain/models/FlashAssessmentSuccessRateHandler.js';
 import { Readable } from 'stream';
 
 async function simulateFlashAssessmentScenario(
