+++ api/src/certification/flash-certification/routes.js
@@ -1,0 +1,5 @@
+import * as scenarioSimulator from './application/scenario-simulator-route.js';
+
+const scenarioSimulatorRoutes = [scenarioSimulator];
+
+export { scenarioSimulatorRoutes };
