+++ api/src/certification/shared/domain/usecases/index.js
@@ -6,6 +6,7 @@ import * as badgeRepository from '../../../../../lib/infrastructure/repositories
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
 import * as certificationChallengeLiveAlertRepository from '../../../session/infrastructure/repositories/certification-challenge-live-alert-repository.js';
 import * as certificationIssueReportRepository from '../../../shared/infrastructure/repositories/certification-issue-report-repository.js';
+import * as challengeRepository from '../../../shared/infrastructure/repositories/challenge-repository.js';
 import * as complementaryCertificationBadgesRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-badge-repository.js';
 import * as complementaryCertificationRepository from '../../../../../lib/infrastructure/repositories/complementary-certification-repository.js';
 import * as complementaryCertificationForTargetProfileAttachmentRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-for-target-profile-attachment-repository.js';
@@ -17,6 +18,7 @@ import * as sessionCodeService from '../../../session/domain/services/session-co
 import * as sessionValidator from '../../../session/domain/validators/session-validator.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
 import * as sessionRepository from '../../../session/infrastructure/repositories/session-repository.js';
+import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
 import * as sessionForAttendanceSheetRepository from '../../../session/infrastructure/repositories/session-for-attendance-sheet-repository.js';
 import * as attendanceSheetPdfUtils from '../../../session/infrastructure/utils/pdf/attendance-sheet-pdf.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
@@ -29,10 +31,12 @@ const dependencies = {
   certificationCenterRepository,
   certificationChallengeLiveAlertRepository,
   certificationIssueReportRepository,
+  challengeRepository,
   complementaryCertificationBadgesRepository,
   complementaryCertificationRepository,
   complementaryCertificationForTargetProfileAttachmentRepository,
   complementaryCertificationTargetProfileHistoryRepository,
+  flashAlgorithmService,
   issueReportCategoryRepository,
   organizationRepository,
   sessionCodeService,
@@ -52,6 +56,9 @@ const usecasesWithoutInjectedDependencies = {
   })),
   ...(await importNamedExportsFromDirectory({
     path: join(path, '../../../complementary-certification/domain/usecases/'),
+  })),
+  ...(await importNamedExportsFromDirectory({
+    path: join(path, '../../../flash-certification/domain/usecases/'),
   })),
 };
