+++ api/src/school/shared/usecases/index.js
@@ -4,7 +4,7 @@ import { dirname, join } from 'node:path';
 import * as activityAnswerRepository from '../../../../lib/infrastructure/repositories/activity-answer-repository.js';
 import * as activityRepository from '../../../../lib/infrastructure/repositories/activity-repository.js';
 import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as challengeRepository from '../../../../lib/infrastructure/repositories/challenge-repository.js';
+import * as challengeRepository from '../../../certification/shared/infrastructure/repositories/challenge-repository.js';
 import * as schoolRepository from '../../../../src/school/infrastructure/repositories/school-repository.js';
 import * as organizationLearnersRepository from '../../../../src/school/infrastructure/repositories/organization-learners-repository.js';
