+++ api/tests/certification/flash-certification/integration/application/scenario-simulator-controller_test.js
@@ -1,10 +1,10 @@
-import { expect, sinon, HttpTestServer, domainBuilder, parseJsonStream } from '../../../test-helper.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import * as moduleUnderTest from '../../../../lib/application/scenarios-simulator/index.js';
-import { random } from '../../../../lib/infrastructure/utils/random.js';
-import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
-import { pickAnswerStatusService } from '../../../../lib/domain/services/pick-answer-status-service.js';
-import { pickChallengeService } from '../../../../lib/domain/services/pick-challenge-service.js';
+import { expect, sinon, HttpTestServer, domainBuilder, parseJsonStream } from '../../../../test-helper.js';
+import { usecases } from '../../../../../src/certification/shared/domain/usecases/index.js';
+import * as moduleUnderTest from '../../../../../src/certification/flash-certification/application/scenario-simulator-route.js';
+import { random } from '../../../../../lib/infrastructure/utils/random.js';
+import { securityPreHandlers } from '../../../../../lib/application/security-pre-handlers.js';
+import { pickAnswerStatusService } from '../../../../../lib/domain/services/pick-answer-status-service.js';
+import { pickChallengeService } from '../../../../../lib/domain/services/pick-challenge-service.js';
 
 describe('Integration | Application | scenario-simulator-controller', function () {
   let httpTestServer;
